requirements.txt
django>=4.2,<5.0
djangorestframework
celery[redis]
psycopg2-binary
redis
django-environ